Extending AppTest with adjustments on reduced model
AppTest included tests only as comment blocks because the reduced learning model included less word vectors. Now, the amount of tests is equal to those that were used with the whole model. They cannot be the same because we get test failures if we would use the old tests on the reduced learning model. Please, write tests for larger model yourself.